pyproject.toml
[project]
name = "fambot-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "matplotlib>=3.10.8",
    "pandas>=3.0.2",
    "scikit-learn>=1.8.0",
    "xgboost>=3.2.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "firebase-admin>=6.5.0",
    "pydantic>=2.10.0",
    "email-validator>=2.0.0",
    "joblib>=1.4.0",
    "PyJWT>=2.9.0",
    "python-dotenv>=1.2.2",
]

[project.scripts]
model = "model:main"
api = "fambot_backend.app:run"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["fambot_backend*"]

[tool.setuptools]
py-modules = ["model"]
